# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NU:TONIC Game Server API
  version: 0.1.0
  description: |
    Versioned REST for Kotlin clients under **`/api/v1/*`**.

    **Server URL vs paths:** `servers[0].url` is the deployment **origin only** (no `/api/v1` suffix).
    Every operation path includes the full **`/api/v1/...`** prefix so URL joiners (RFC 3986) resolve
    correctly—do not combine `https://host/api/v1` with path `/health` (that resolves to `https://host/health`).

    **Feature toggles (`GET /api/v1/config` → `features`):** When `community_lb_get` or `community_lb_post`
    is `false`, matching routes return **403** with `application/json` body
    `{"error":"feature_disabled","feature":"<name>"}` (`plans/2026-04-13-product-flags-v1.md`).

    **Community leaderboard POST** is **not** ranked verification: client-supplied scores/distances are
    presentation or lab aggregates only (`rules/05-networking-leaderboard.md`).

    Hand-maintained; keep in sync with `server/src/nutonic_server/` until export automation lands (IMP-011).
servers:
  - url: https://example.invalid
    description: Replace with deployment origin (scheme + host, optional port); paths carry `/api/v1`
...
paths:
...
  /api/v1/debug/session:
    get:
      tags: [debug]
      summary: Validate Bearer JWT (debug)
      operationId: getDebugSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Token OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, session_id]
                properties:
                  ok:
                    type: boolean
                  session_id:
                    type: string
        "401":
          description: Missing or invalid token
  /api/v1/maps/{map_id}/leaderboard:
    get:
      tags: [leaderboard]
      summary: In-memory leaderboard rows (S0 community slice)
      description: |
        Optional community / lab aggregate when `features.community_lb_get` is true.
        When false, **403** `feature_disabled`.
      operationId: getMapLeaderboard
      parameters:
        - name: map_id
          in: path
          required: true
          schema:
            type: string
        - name: tier
          in: query
          required: false
          schema:
            type: string
            enum: [ranked]
          description: |
            When **`ranked`**, return the same server-verified rows as **`GET /api/v1/maps/{map_id}/leaderboard/ranked`**
            (`docs/RANKED-MODE.md` §4). Requires **`features.ranked`**. Omit for the default **community** board (requires **`features.community_lb_get`**).
      responses:
        "200":
          description: Rows for map_id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardRow"
        "403":
          description: features.community_lb_get is false, or features.ranked is false when tier=ranked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDisabledError"
    post:
      tags: [leaderboard]
      summary: Append a dev/lab community leaderboard row (JWT required)
      description: |
        Non-authoritative for non-ranked math; **not** a ranked `submit` path.
        Duplicate **`Idempotency-Key`** for the same `map_id` returns the first stored row (HTTP 200).
      operationId: postMapLeaderboard
      security:
        - bearerAuth: []
      parameters:
        - name: map_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: "Recommended on writes; server dedupes when present (rules/05)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderboardPostIn"
      responses:
        "200":
          description: Stored row (or prior row when Idempotency-Key repeats)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardRow"
        "401":
          description: Missing or invalid token
        "403":
          description: features.community_lb_post is false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDisabledError"
  /api/v1/maps/{map_id}/leaderboard/ranked:
    get:
      tags: [leaderboard, ranked]
      summary: Server-verified ranked scores for a map (aggregate)
      description: |
        Rows produced only from successful **`POST /api/v1/ranked/rounds/{round_id}/submit`** (guess-only verification).
        Separate from community **`GET /api/v1/maps/{map_id}/leaderboard`** (`docs/RANKED-MODE.md` §4).
        When **`features.ranked`** is false, **403** `feature_disabled`.
      operationId: getMapLeaderboardRanked
      parameters:
        - name: map_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verified ranked rows, highest score first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardRow"
        "403":
          description: features.ranked is false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDisabledError"
...
  /api/v1/cache/manifest:
    get:
      tags: [cache]
      summary: Hydration manifest (catalog + content_version)
      description: |
        Client cache anchor per **`rules/13`** (**IMP-080**). Response includes **`ETag`**; send
        **`If-None-Match`** to receive **304** when unchanged. Large bundle bytes stay on CDN/workers—this
        JSON is the index only (`plans/2026-04-07-complete-implementation-architecture.md` §9 **S3**).

        **Public spoiler hygiene:** the reference server may omit **`locations`** and **`ai_guesses`**
        unless **`NUTONIC_EXPOSE_MANIFEST_ROUND_TRUTH=true`** is set (dev/CI). Treat world-readable
        manifests as non-secret; never ship pre-round golden truth to untrusted clients in production.
      operationId: getCacheManifest
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
          description: Prior **ETag** from a successful **200** response
      responses:
        "200":
          description: Manifest body (revalidate or first fetch)
          headers:
            ETag:
              required: true
              schema:
                type: string
              description: Weak entity tag for this canonical JSON payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheManifest"
        "304":
          description: Not modified — reuse cached body + metadata
          headers:
            ETag:
              required: true
              schema:
                type: string
  /api/v1/bundles/{bundle_id}:
    get:
      tags: [cache]
      summary: Download versioned still bytes (IMP-081)
      operationId: getBundleStill
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: JPEG still bytes
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Unknown bundle id
  /api/v1/maps/{map_id}/guesses/record:
    post:
      tags: [telemetry]
      summary: Record a non-ranked guess (non-authoritative telemetry)
      operationId: postMapGuessRecord
      security:
        - bearerAuth: []
      parameters:
        - name: map_id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuessRecordIn"
      responses:
        "200":
          description: Row stored (or idempotent replay)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuessRecordOut"
        "401":
          description: Missing or invalid token
        "403":
          description: features.guesses_record is false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDisabledError"
  /api/v1/ranked/rounds/start:
    post:
      tags: [ranked]
      summary: Start a server-secret ranked round (IMP-090)
      operationId: postRankedRoundStart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RankedRoundStartIn"
      responses:
        "200":
          description: Round id + short-lived round_ticket + redacted clue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankedRoundStartOut"
        "401":
          description: Missing or invalid token
        "403":
          description: features.ranked is false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDisabledError"
        "404":
          description: Unknown map_id for ranked catalog
  /api/v1/ranked/rounds/{round_id}/forfeit-ranked-integrity:
    post:
      tags: [ranked]
      summary: Forfeit ranked verified placement before submit (IMP-091)
      operationId: postRankedRoundForfeitIntegrity
      security:
        - bearerAuth: []
      parameters:
        - name: round_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RankedForfeitIn"
      responses:
        "200":
          description: Round marked forfeited; round_ticket is no longer valid for verified submit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankedForfeitOut"
        "401":
          description: Missing session token
        "403":
          description: features.ranked off, or session does not own this round
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureDisabledError"
        "404":
          description: Unknown round_id
        "409":
          description: Round not open (already submitted or forfeited)
...
  /api/v1/pro/jobs:
    post:
      tags: [pro]
      summary: Create PRO materialization job (stub control plane)
      operationId: postProJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProJobCreateIn"
      responses:
        "200":
          description: Job accepted (stub)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProJobCreateOut"
        "403":
          description: features.pro_jobs is false
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    FeatureDisabledError:
      type: object
      required: [error, feature]
      properties:
        error:
          type: string
          example: feature_disabled
        feature:
          type: string
          example: community_lb_get
...
    LeaderboardRow:
      type: object
      required: [display_handle, player_role, score_points]
      properties:
        display_handle:
          type: string
          maxLength: 64
        player_role:
          type: string
        score_points:
          type: integer
        distance_km:
          type: number
          format: float
          nullable: true
...
    GuessRecordOut:
      type: object
      required: [id, recorded]
      properties:
        id:
          type: integer
        recorded:
          type: boolean
    RankedRoundStartIn:
      type: object
      required: [map_id]
      properties:
        map_id:
          type: string
...